+++ package.json
@@ -8,8 +8,8 @@
 		"cli":"node dist/cli/index.js",
 		"test": "jest --config=tests/jest.config.cjs",
 		"test:coverage": "npm run test -- --collectCoverage && open coverage/lcov-report/index.html",
-		"dev": "webpack --config webpack.config.cjs --watch",
-		"build": "cross-env NODE_ENV=production webpack --config webpack.config.cjs",
+		"extension:dev": "npm run dev --workspace=@cookie-analysis-tool/extension",
+		"extension:build": "npm run build --workspace=@cookie-analysis-tool/extension",
 		"serve": "webpack serve",
 		"lint": "npm-run-all --parallel lint:*",
 		"lint:js": "eslint .",
