+++ packages/extension/babel.config.cjs
@@ -14,13 +14,6 @@
  * limitations under the License.
  */
 
-const path = require('path');
-
 module.exports = {
-  plugins: {
-    tailwindcss: {
-      config: path.join(__dirname, './tailwind.config.cjs'),
-    },
-    autoprefixer: {},
-  },
+  extends: '../../babel.config.cjs',
 };
